Caching the root domain in the mixed callstack implementation to write out on every domain reload.
When a domain reload happens the mixed callstack file gets deleted and rewritten with the jitted functions. Since the root domain is not reloaded it is no longer included in the mixed callstack file and therefore no longer resolves in the callstack window. This change caches the root callstack frames and writes them out after a domain unload occurs to ensure that these frames will display properly.